third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-css-resources-expected.txt

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-resources.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/css/resources/cors-ahem.php': this content should also be loaded over HTTPS.

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-resources.html' was loaded over HTTPS, but displayed insecure content from 'http://127.0.0.1:8080/resources/square20.png': this content should also be loaded over HTTPS.

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-resources.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/css/resources/cors-ahem.php': this content should also be loaded over HTTPS.

This test opens a window that loads a style sheet which fetches an insecure background image and an insecure web font. We should trigger a mixed content callback because the main frame in the window is HTTPS but is running insecure content.
